feat: Adds STX & STY
