+++ test/opcodes/CMakeLists.txt
@@ -1,0 +1,30 @@
+# 1. Declare the test file
+set(Tests
+  sta_test
+  stx_test
+  sty_test)
+
+###################################################################
+
+foreach( tst ${Tests} )
+  message(STATUS "test=${tst}")
+  add_executable(
+    ${tst}
+    ${tst}.c)
+
+  target_link_libraries(
+    ${tst}
+    unity
+    cpu)
+
+  target_include_directories(
+    ${tst} PUBLIC
+    ${PROJECT_SOURCE_DIR}/test/unity)
+
+  set_target_properties(
+    ${tst}
+    PROPERTIES
+    RUNTIME_OUTPUT_DIRECTORY "${TEST_OUTPUT_PATH}")
+
+  add_test(${tst} "${TEST_OUTPUT_PATH}/${tst}")
+endforeach()
